Refactor the code to make it more testable
